+++ geostore/tiles/mixins.py
@@ -2,7 +2,7 @@ import json
 from urllib.parse import unquote, urljoin
 
 from django.core.cache import cache
-from django.http import HttpResponse, HttpResponseNotFound
+from django.http import HttpResponse
 from django.urls import reverse
 
 from ..models import Feature, Layer
@@ -120,41 +120,38 @@ class AbstractTileJsonMixin:
         self.object = self.get_object()
         last_update = self.get_last_update()
 
-        if last_update:
-            cache_key = f'tilejson-{self.object.name}'
-            version = int(last_update.timestamp())
-            tilejson_data = cache.get(cache_key, version=version)
+        cache_key = f'tilejson-{self.object.name}'
+        version = int(last_update.timestamp())
+        tilejson_data = cache.get(cache_key, version=version)
 
-            if tilejson_data is None:
-                tilejson_data = json.dumps(self.get_tilejson())
-                cache.set(cache_key, tilejson_data, version=version)
+        if tilejson_data is None:
+            tilejson_data = json.dumps(self.get_tilejson())
+            cache.set(cache_key, tilejson_data, version=version)
 
-            response_kwargs.setdefault('content_type', self.content_type)
-            return self.response_class(
-                content=tilejson_data,
-                **response_kwargs,
-            )
-
-        return HttpResponseNotFound()
+        response_kwargs.setdefault('content_type', self.content_type)
+        return self.response_class(
+            content=tilejson_data,
+            **response_kwargs,
+        )
 
 
 class TileJsonMixin(AbstractTileJsonMixin):
-
     def get_tile_path(self):
         return reverse("geostore:layer-tiles-pattern", args=[self.object.pk])
 
     @property
     def layers(self):
-        return [self.object]
+        # keep a qs result here
+        return type(self.object).objects.filter(pk=self.object.pk)
 
 
 class MultipleTileJsonMixin(AbstractTileJsonMixin):
-
     def get_tile_path(self):
         return reverse("geostore:group-tiles-pattern", args=[self.object.slug])
 
     @property
     def layers(self):
+        # keep a qs result here
         return self.object.layers.all()