+++ geostore/tiles/mixins.py
@@ -4,6 +4,8 @@ from urllib.parse import unquote, urljoin
 from django.core.cache import cache
 from django.http import HttpResponse
 from django.urls import reverse
+from django.utils.encoding import escape_uri_path
+from django.utils.html import escape
 
 from ..models import Feature, Layer
 from .. import settings as app_settings
@@ -19,8 +21,9 @@ class AbstractTileJsonMixin:
         layer_settings = layer.layer_settings_with_default(*args)
         if 'link' in layer_settings:
             return '<a href="{0}"/>{1}</a>'.format(
-                layer_settings['link'].replace('"', '&quot;'),
-                layer_settings['name'].replace('"', '&quot;'))
+                escape_uri_path(layer_settings['link']),
+                escape(layer_settings['name'])
+            )
         return layer_settings
 
     @property
@@ -46,22 +49,26 @@ class AbstractTileJsonMixin:
                 for l in self.layers
             ]))
 
-    def get_attribution(self):
-        return ','.join(set(
-            [
-                l.layer_settings_with_default('metadata', 'attribution')
-                for l in self.layers if l.layer_settings_with_default('metadata', 'attribution')
-            ])) or None
-
-    def get_description(self):
+    def _join_group_settings_link(self, layers, *args):
         return ','.join(set([
-            a
-            for a in [
-                layer.layer_settings_with_default('metadata', 'description')
-                for layer in self.layers
-            ]
+            a if 'link' not in a else
+            '<a href="{0}"/>{1}</a>'.format(escape_uri_path(a['link']), escape(a['name']))
+            for a in [layer.layer_settings_with_default(*args) for layer in self.layers]
             if a
         ])) or None
+
+    def _join_group_settings_string(self, layers, *args):
+        return ','.join(set([
+            a
+            for a in [layer.layer_settings_with_default(*args) for layer in self.layers]
+            if a
+        ])) or None
+
+    def get_attribution(self):
+        return self._join_group_settings_link(self.layers, 'metadata', 'attribution')
+
+    def get_description(self):
+        return self._join_group_settings_string(self.layers, 'metadata', 'description')
 
     @staticmethod
     def layer_fields(layer):