+++ CHANGES.md
@@ -10,6 +10,7 @@ CHANGELOG
 * Fix bug with shapefile export on geometry defined layer.
 * Fix group's tiles URLs in tilejson
 * Fix tilejson when Layer has no Feature
+* Return a tilejson even if it has no feature
 
 
 0.3.4      (2019-09-26)