+++ CHANGES.md
@@ -8,7 +8,7 @@ CHANGELOG
 ### Fixes
 
 * Fix deprecation warning : "ST_Line_Substring signature was deprecated in 2.1.0. Please use ST_LineSubstring"
-
+* Fix tile generation when no feature is present in the layer
 
 0.3.5      (2019-10-03)
 -----------------------